Moved pagination links creation to utility file
